Move audit package under resources
