+++ discord/audit/entry_emoji.go
