+++ discord/audit/entry_role.go
