+++ discord/audit/parse.go
